[refactor]: Card 컴포넌트 분리
