+++ src/components/list/List.tsx
@@ -1,37 +1,24 @@
 import React from 'react';
 import { SectionType } from '../../types/enums';
-import { useAppDispatch, useAppSelector } from '../../hooks/useRTK';
-import Button from '../common/button/Button';
-import { updateTodo } from '../../store/slice/todoSlice';
+import { useAppSelector } from '../../hooks/useRTK';
+import Card from '../card/Card';
 
 interface Props {
   sectionInfo: SectionType;
 }
 export default function List({ sectionInfo }: Props) {
-  const dispatch = useAppDispatch();
-
   const todos = useAppSelector((state) => state.todoSlice.todos).filter(
     (todo) => !sectionInfo === todo.isDone,
   );
 
-  const handleUpdateTodo = (id: number) => {
-    dispatch(updateTodo(id));
-  };
+  const TODO_CARD = todos.map((todo) => <Card key={todo.id} todo={todo} />);
+
   return (
     <section>
       <h2>{sectionInfo ? 'Done' : 'Not Done'}</h2>
       <ul>
-        {todos.map((todo) => (
-          <li key={todo.id}>
-            <h2>{todo.title}</h2>
-            <p>{todo.contents}</p>
-            <Button
-              isSubmit={false}
-              text={sectionInfo ? 'Cancel Done' : 'Done'}
-              handler={() => handleUpdateTodo(todo.id)}
-            />
-          </li>
-        ))}
+        {/* TODO: 리스트 아이템 컴포넌트 분리, 삭제버튼 및 로직 추가, 업데이트 로직 추가 */}
+        {TODO_CARD}
       </ul>
     </section>
   );
